clean up code in Expertise.js
