+++ src/text/expertise.js
@@ -1,0 +1,31 @@
+import React from 'react';
+import {
+  FaRandom,
+  FaGlobeAmericas,
+  FaComments,
+  FaFileAlt,
+  FaStar,
+} from 'react-icons/fa';
+
+export const expertise = [
+  {
+    text: ['Able to quickly', 'grasp complex,', 'technical', 'solutions'],
+    icon: <FaRandom size={36} />,
+  },
+  {
+    text: ['Extensive', 'experience in', 'international', 'business'],
+    icon: <FaGlobeAmericas size={36} />,
+  },
+  {
+    text: ['Translate', 'tech talk into', 'non-technical', 'language'],
+    icon: <FaComments size={36} />,
+  },
+  {
+    text: ['Can see past', 'complexity to', 'bring out the', 'best'],
+    icon: <FaStar size={36} />,
+  },
+  {
+    text: ['Both', 'copywriting', 'and layout', 'design'],
+    icon: <FaFileAlt size={36} />,
+  },
+];
